bug fixing, changing a default rule
fixing incompatibility bug for 2023.10.1.1 #9
changing default value of Last Screen Position And Size to false so it needs to be enabled manually now! This is to avoid a bug in Burp Suite when Burp window is moved before all extensions are loaded which can make the repeater sub-tabs' layout disappear!
Montoya API now uses 2023.10 rather than 2023.10.1